changed type of some fields on Long and Bigdecimal

+++ src/main/java/com.accenture.flowershop/be/business/UserBusinessService.java
@@ -4,11 +4,13 @@ import com.accenture.flowershop.be.entity.user.Customer;
 import com.accenture.flowershop.fe.dto.CustomerDTO;
 import com.accenture.flowershop.fe.enums.customer.UserShop;
 
+import java.math.BigDecimal;
+
 public interface UserBusinessService {
 
     void DeleteUserInMap(String loginUser);
     Customer logIn(String user, String password);
-    Customer register(String login, String password, String surname, String name, String patronymic, String address, double cashBalance, int discount, UserShop userRole);
+    Customer register(String login, String password, String surname, String name, String patronymic, String address, BigDecimal cashBalance, BigDecimal discount, UserShop userRole);
     Customer saveUser(Customer customer);
     int findIdUser(String login);
 }
